MedC/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-ccs811==1.3.16
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-register==1.10.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.3.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.3.5
attrs==23.2.0
auto-py-to-exe==2.42.0
Babel==2.14.0
beeware==0.3.0
Bext==0.1.1
binaryornot==0.4.4
binho-host-adapter==0.1.6
blinker==1.9.0
bottle==0.12.25
bottle-websocket==0.2.9
briefcase==0.3.20
build==1.2.2.post1
buildozer==1.5.0
cachetools==5.5.0
catboost==1.2.7
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.5
comtypes==1.2.1
configparser==4.0.2
contourpy==1.3.0
cookiecutter==2.6.0
coverage==7.6.8
CTkMessagebox==2.5
customtkinter==5.2.2
cycler==0.12.1
Cython==3.0.11
darkdetect==0.8.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2025.2.0
dask-jobqueue==0.9.0
DateTime==5.5
diagrams==0.24.1
dill==0.3.9
distlib==0.3.8
distributed==2025.2.0
Django==5.0.1
dnspython==2.7.0
docutils==0.21.2
Eel==0.16.0
email_sender==1.2
email_validator==2.2.0
et_xmlfile==2.0.0
evtx==0.8.3
fastapi==0.115.0
filelock==3.13.1
firebird-base==1.8.0
firebird-driver==1.10.10
firebirdsql==1.3.5
flake8==7.1.1
Flask==3.1.0
flask-form==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.10.0
future==0.18.3
geographiclib==2.0
geopy==2.4.1
gevent==23.9.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.2.1
GPUtil==1.4.0
graphviz==0.20.3
greenlet==3.0.3
gspread==6.1.3
h11==0.14.0
hexdump==3.3
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.7
image==1.5.33
imageio==2.37.0
IMAPClient==3.0.1
importlib_metadata==8.5.0
iniconfig==2.0.0
iso8601==2.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.2.0
kiwisolver==1.4.7
lazy_loader==0.4
locket==1.0.0
lxml==5.3.0
mando==0.7.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
more-itertools==5.0.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==23.2
pandas==2.2.3
partd==1.4.2
patsy==1.0.1
pefile==2023.2.7
pexpect==4.9.0
pillow==10.2.0
plantcv==4.6
platformdirs==4.1.0
playsound==1.3.0
plotly==5.24.1
pluggy==1.5.0
plumbum==1.8.2
protobuf==6.30.2
psutil==5.9.7
psycopg2==2.9.9
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydocstyle==6.3.0
pyflakes==3.2.0
pyftdi==0.56.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pylint==3.3.1
pymodbus==3.6.5
PyMuPDF==1.23.14
PyMuPDFb==1.23.9
pyparsing==2.4.7
pypiwin32==223
pyproject-toml==0.0.10
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.7
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
pyserial==3.5
pytest==8.3.3
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-evtx==0.7.4
python-multipart==0.0.12
python-slugify==8.0.4
python-telegram-bot==21.6
pyttsx3==2.90
pytz==2024.1
pyusb==1.2.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.2
radon==6.0.1
referencing==0.32.1
regex==2024.11.6
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.17.1
rsa==4.9
safetensors==0.5.3
schedule==1.2.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
serial==0.0.97
setuptools==75.8.0
sh==2.0.6
six==1.16.0
smmap==5.0.1
smsru_api==0.1.7
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sqlalchemy-firebird==2.1
sqlparse==0.4.4
starlette==0.38.6
statsmodels==0.14.4
sympy==1.13.1
tblib==3.0.0
tenacity==9.0.0
text-unidecode==1.3
threaded==4.2.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tkcalendar==1.6.1
tkPDFViewer==0.1
tokenizers==0.21.1
toml==0.10.2
tomli_w==1.1.0
tomlkit==0.13.2
toolz==1.0.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.66.1
transformers==4.49.0
ttkbootstrap==1.10.1
ttkthemes==3.2.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.1
uuid6==2024.7.10
uvicorn==0.31.0
virtualenv==20.25.0
vl-convert-python==1.7.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.42.0
whichcraft==0.6.1
winapps==0.2.0
WMI==1.5.1
WTForms==3.1.2
xarray==2025.1.2
xgboost==3.0.0
xmltodict==0.14.2
yarl==1.13.1
zict==3.0.0
zipp==3.21.0
zope.event==5.0
zope.interface==6.1
